requirements.txt
asgiref==3.2.7
bleach==3.1.5
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
decorator==4.4.2
Django==3.0.7
django-cors-headers==3.4.0
django-extensions==2.2.9
django-filter==1.1.0
django-haystack==2.8.1
django-rest-swagger==2.2.0
djangorestframework==3.8.2
docutils==0.16
idna==2.9
imageio==2.8.0
importlib-metadata==1.6.1
imutils==0.5.3
itypes==1.2.0
Jinja2==2.11.2
jsonfield==3.1.0
keyring==21.2.1
kiwisolver==1.2.0
lxml==4.5.0
Markdown==2.6.6
MarkupSafe==1.1.1
matplotlib==3.2.1
networkx==2.4
numpy==1.18.4
openapi-codec==1.3.2
opencv-python==4.2.0.34
opencv-python-headless==4.2.0.34
packaging==20.4
pdfminer==20191125
Pillow==7.1.2
pkginfo==1.5.0.1
protobuf==3.11.4
pycryptodome==3.9.7
Pygments==2.6.1
pyparsing==2.4.7
PyQt5==5.14.2
PyQt5-sip==12.7.2
pytesseract==0.3.4
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
readme-renderer==26.0
requests==2.24.0
requests-toolbelt==0.9.1
scikit-image==0.17.2
scipy==1.4.1
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
tifffile==2020.6.3
tqdm==4.46.1
twine==3.1.1
uritemplate==3.0.1
urllib3==1.25.9
webencodings==0.5.1
Whoosh==2.7.4
zipp==3.1.0
